refactor: Moved recipe controller to separate directory
