+++ test/recipe_scripts_tests/eat_this_test.dart
@@ -3,7 +3,7 @@ import 'package:html/dom.dart';
 import 'package:ingredient_collector/l10n/locale_keys.g.dart';
 import 'package:ingredient_collector/src/models/ingredient.dart';
 import 'package:ingredient_collector/src/models/recipe_parsing_job.dart';
-import 'package:ingredient_collector/src/recipe_controller.dart';
+import 'package:ingredient_collector/src/recipe_controller/recipe_controller.dart';
 import 'package:ingredient_collector/src/recipe_scripts/recipe_parser.dart'
     show EatThisParser;
